requirements.txt
openai-whisper==20231117
yt-dlp==2024.07.09
webvtt-py==0.4.6
transformers==4.44.2
torch==2.4.0
ffmpeg-python==0.2.0